Rotation functions and commends added
This is the code for the base autonomous movement for the Tekbot. Currently the autonomous movement is not fully fleshed out, but the basic functions for movement exist.  Comments were added into this iteration of robotMovement for clarity. The functions rotateRight and rotateLeft were updated to be functional with current understanding of Tekbot orientation. Testing is required.