Ultrasonic Sensor and touch sensor capabilities added. Autonomous Maze navigation also added.
This particular update has a lot in it, however the functions are pretty self explanatory. A few of the functions include isTouchSensorPressed howFarIsWall, and navigateMaze. The HIGH and LOW inputs on the motors were also adjusted as they were set incorrectly before. Although touch sensor capabilities are currently possible, they have yet to be implemented into the Autonomous Maze navigation.